requirements.txt
requests
python-whois
tldextract
python-nmap
